handle missing !plugins arguments
Prior to this commit, if a user posted "!plugins" with no other
arguments to helga, helga would error with a strack trace. We were
referencing the args list's first element without verifying that the
args list has a first element.

Update the "!plugins" command so that it doesn't crash and calls the
same code path as "!plugins list" instead. The assumption is that this
is generally what the user would want.